Removed to warnings related to wrong type in for-loop
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@13745 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/core/Runner.hh
@@ -124,7 +124,7 @@ namespace karabo {
                     resolveTokens(args, resolved);
 
                     std::vector<std::string> tokenList;
-                    for (int i = 0; i < resolved.size(); ++i) {
+                    for (size_t i = 0; i < resolved.size(); ++i) {
                         parseToken("", resolved[i], tokenList);
                     }
 
@@ -194,7 +194,7 @@ namespace karabo {
                         vector<string> args, resolved;
                         boost::split(args, line, boost::is_any_of(" "));
                         resolveTokens(args, resolved);
-                        for (int i = 0; i < resolved.size(); ++i) {
+                        for (size_t i = 0; i < resolved.size(); ++i) {
                             parseToken(key, resolved[i], tokenList);
                         }
                     } else {
